+++ CHANGELOG.md
@@ -24,6 +24,11 @@
   
 - `ui-components`
   - [#2553](https://github.com/wix-incubator/rich-content/pull/2553) moved labeledToggle component into ui-components
+  
+### :rocket: New Feature
+- `ricos-content`
+  - [#2550](https://github.com/wix/ricos/pull/2550) media extractor for SEO
+
 </details>
 <hr/>